Results after Resolver Revision
