+++ src/graphql/schema/uuid/comment/resolvers.ts
@@ -1,8 +1,10 @@
+import { createUuidResolvers } from '../abstract-uuid'
 import { UserPayload } from '../user'
 import { CommentResolvers } from './types'
 
 export const resolvers: CommentResolvers = {
   Comment: {
+    ...createUuidResolvers(),
     createdAt(comment) {
       return comment.date.toString()
     },
